automate release versioning, move create_logger to helpers
