+++ Pipfile
@@ -19,6 +19,3 @@ coveralls = "*"
 sphinx_rtd_theme = "*"
 mypy = "*"
 black = "==19.10b0"
-
-[requires]
-python_version = "3.5"
