+++ pytube/__init__.py
@@ -9,7 +9,7 @@ __author__ = "Nick Ficano, Harold Martin"
 __license__ = "MIT License"
 __copyright__ = "Copyright 2019 Nick Ficano"
 
-from pytube.logging import create_logger
+from pytube.helpers import create_logger
 from pytube.query import CaptionQuery
 from pytube.query import StreamQuery
 from pytube.streams import Stream
