+++ pytube/version.py
@@ -1,1 +1,4 @@
 __version__ = "9.5.6"
+
+if __name__ == "__main__":
+    print(__version__)
